rename s3 classes to be reuse-able
